+++ lib/commands/code/src/code_command_base.dart
@@ -1,0 +1,6 @@
+import 'package:alex/runner/alex_command.dart';
+
+/// Base command for localization feature.
+abstract class CodeCommandBase extends AlexCommand {
+  CodeCommandBase(String name, String description) : super(name, description);
+}