+++ lib/runner/alex_command.dart
@@ -1,5 +1,9 @@
+import 'dart:io';
+
+import 'package:alex/src/exception/run_exception.dart';
 import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
+import 'package:meta/meta.dart';
 
 /// Базовый класс команды.
 abstract class AlexCommand extends Command<int> {
@@ -7,7 +11,7 @@ abstract class AlexCommand extends Command<int> {
   final String _description;
   final ArgParser _argParser = ArgParser(
     allowTrailingOptions: false,
-  );
+  )..addFlag('verbose', help: 'Show additional diagnostic info');
 
   AlexCommand(this._name, this._description);
 
@@ -19,4 +23,75 @@ abstract class AlexCommand extends Command<int> {
 
   @override
   String get description => _description;
+
+  /// Prints message if verbose flag is on.
+  @protected
+  void printVerbose(String message) {
+    if (argResults['verbose'] as bool) print(message);
+  }
+
+  /// Prints some info message in output.
+  @protected
+  void printInfo(String message) => print(message);
+
+  /// Prints error message in error output.
+  @protected
+  void printError(String message) => stderr.writeln(message);
+
+  /// Prints 0 code and prints a success message if provided.
+  @protected
+  int success({String message}) {
+    if (message != null) printInfo(message);
+    return 0;
+  }
+
+  /// Returns error code and prints a error message if provided.
+  @protected
+  int error(int code, {String message}) {
+    if (message != null) printError(message);
+    return code;
+  }
+
+  /// Returns error code by exception.
+  @protected
+  int errorBy(RunException exception) {
+    assert(exception != null);
+    return error(exception.exitCode, message: exception.message);
+  }
+
+  /// Runs `flutter pub run` command.
+  @protected
+  Future<ProcessResult> runPub(String cmd, List<String> arguments) async {
+    final executable = 'flutter';
+    final args = ['pub', 'run', cmd, ...arguments];
+
+    printVerbose('Run: $executable ${args.join(" ")}');
+
+    return Process.run(executable, args);
+  }
+
+  @protected
+  Future<ProcessResult> runPubOrFail(String cmd, List<String> arguments,
+      {bool printStdOut = true}) async {
+    return runOrFail(() => runPub(cmd, arguments), printStdOut: printStdOut);
+  }
+
+  @protected
+  Future<ProcessResult> runOrFail(Future<ProcessResult> Function() run,
+      {bool printStdOut = true}) async {
+    final res = await run();
+
+    // TODO: prints out during invocation
+
+    if (res.exitCode != 0) {
+      throw RunException(res.exitCode, res.stderr.toString());
+    }
+
+    final runOut = res.stdout?.toString();
+    if (printStdOut && runOut != null && runOut.isNotEmpty) {
+      printInfo(res.stdout.toString());
+    }
+
+    return res;
+  }
 }