+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 1.0.0
+## 0.2.0-dev.1 - 2020-09-03
 
-- Initial version, created by Stagehand
+- Import translations from xml.
+
+## 0.1.0-dev.2 - 2020-08-14
+
+- Working with localization: extract, generate and export to xml.